playbook-agent/requirements.txt
llama-api-python
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
ansible>=8.0.0
ansible-lint>=6.0.0
paramiko>=3.0.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
httpx>=0.24.0